House Price Prediction
In this project, a linear regression model was used in predicting the sales prices of various houses given their house ids and so many features. The goal of the project is to predict the sales price for each house. For each Id in the test set, I must predict the value of the SalePrice variable.